requirements.txt
requests
kafka-python
aiohttp
pytest
psycopg2-binary
